requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
spacy>=3.7.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
textstat>=0.7.0
nltk>=3.8.0
transformers>=4.30.0
sentence-transformers>=2.2.0
